LSP context: extract LSP commands mocking into a separate file
